\name{zigguserHeaderRcpp-package}
\alias{zigguserHeaderRcpp-package}
\alias{zigguserHeaderRcpp}
\docType{package}
\title{\packageTitle{zigguserHeaderRcpp}}
\description{\packageDescription{zigguserHeaderRcpp}}
\section{Package Content}{\packageIndices{zigguserHeaderRcpp}}
\author{\packageAuthor{zigguserHeaderRcpp}}
\section{Maintainer}{\packageMaintainer{zigguserHeaderRcpp}}
\keyword{package}
